move helpers toslashelper file
